refactor: update documentation and switch to anonymous components for MACI non-core circuits
