+++ circuits/circom/privToPubKey.circom
@@ -1,31 +1,28 @@
 pragma circom 2.0.0;
 
-// circomlib imports
+// from circomlib.
 include "./bitify.circom";
 include "./escalarmulfix.circom";
 
-// convert a private key to a public key
-// @note the basepoint is the base point of the baby jubjub curve
+/**
+ * Converts a private key to a public key on the BabyJubJub curve.
+ * The input private key needs to be hashed and then pruned before.
+ */
 template PrivToPubKey() {
-    // Needs to be hashed, and then pruned before supplying it to the circuit
-    signal input privKey;
-    signal output pubKey[2];
-
-    // convert the private key to bits
-    component privBits = Num2Bits(253);
-    privBits.in <== privKey;
-
+    // The base point of the BabyJubJub curve.
     var BASE8[2] = [
         5299619240641551281634865583518297030282874472190772894086521144482721001553,
         16950150798460657717958625567821834550301663161624707787222815936182638968203
     ];
 
-    // perform scalar multiplication with the basepoint
-    component mulFix = EscalarMulFix(253, BASE8);
-    for (var i = 0; i < 253; i++) {
-        mulFix.e[i] <== privBits.out[i];
-    }
+    signal input privKey;
+    signal output pubKey[2];
 
-    pubKey[0] <== mulFix.out[0];
-    pubKey[1] <== mulFix.out[1];
+    // Convert the private key to bits.
+    var privBits[253] = Num2Bits(253)(privKey);
+
+    // Perform scalar multiplication with the basepoint.
+    var mulFix[2] = EscalarMulFix(253, BASE8)(privBits);
+
+    pubKey <== mulFix;
 }
