+++ circuits/circom/processMessages.circom
@@ -8,8 +8,10 @@ include "./hashers.circom";
 include "./messageToCommand.circom";
 include "./privToPubKey.circom";
 include "./stateLeafAndBallotTransformer.circom";
-include "./trees/incrementalQuinTree.circom";
-include "./utils.circom";
+include "./iqt.circom";
+// zk-kit imports
+include "./safe-comparators.circom";
+include "./processMessages.circom";
 
 // Proves the correctness of processing a batch of messages.
 template ProcessMessages(
