+++ circuits/circom/test/ProcessMessages_10-2-1-2_test.circom
@@ -1,12 +1,6 @@
+// auto-generated by circomkit
 pragma circom 2.0.0;
 
 include "../processMessages.circom";
 
-/*
-stateTreeDepth,
-msgTreeDepth,
-msgSubTreeDepth
-voteOptionTreeDepth,
-*/
-
-component main {public [inputHash]} = ProcessMessages(10, 2, 1, 2);
+component main {public[inputHash]} = ProcessMessages(10, 2, 1, 2);
