+++ circuits/circom/test/SubsidyPerBatch_10-1-2_test.circom
@@ -1,5 +1,6 @@
+// auto-generated by circomkit
 pragma circom 2.0.0;
 
 include "../subsidy.circom";
 
-component main {public [inputHash]} = SubsidyPerBatch(10, 1, 2);
+component main {public[inputHash]} = SubsidyPerBatch(10, 1, 2);
