+++ circuits/circom/test/TallyVotes_10-1-2_test.circom
@@ -1,5 +1,6 @@
+// auto-generated by circomkit
 pragma circom 2.0.0;
 
 include "../tallyVotes.circom";
 
-component main {public [inputHash]} = TallyVotes(10, 1, 2);
+component main {public[inputHash]} = TallyVotes(10, 1, 2);
